move eq instances to package
